at_learner_core/requirements.txt
absl-py==0.9.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
google-auth==1.10.1
google-auth-oauthlib==0.4.1
grpcio==1.26.0
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.2
Jinja2==2.11.1
joblib==0.14.1
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.2
kiwisolver==1.1.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.1.2.30
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pdoc3==0.7.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.2
plotly==4.5.0
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.11.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.13.0
tensorboard==2.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.3.1
torchfile==0.1.0
torchvision==0.4.2
tornado==6.0.3
tqdm==4.42.0
traitlets==4.3.3
urllib3==1.25.7
visdom==0.1.8.9
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.0
widgetsnbextension==3.5.1
zipp==2.2.0
